+++ frmts/zarr/CMakeLists.txt
@@ -2,7 +2,11 @@ add_gdal_driver(
   TARGET gdal_Zarr
   SOURCES zarr_array.cpp
           zarr_attribute.cpp
+          zarr_dimension.cpp
           zarr_group.cpp
+          zarr_v2_group.cpp
+          zarr_v3_group.cpp
+          zarr_sharedresource.cpp
           zarrdriver.cpp
           STRONG_CXX_WFLAGS
           PLUGIN_CAPABLE
